Start refactoring reservations operator
