+++ reservations/internal/commitments/client_test.go
@@ -1,0 +1,4 @@
+// Copyright 2025 SAP SE
+// SPDX-License-Identifier: Apache-2.0
+
+package commitments
